Пофикшены некоторые баги: 1. Теперь цена не может быть больше 10^12. 2. Теперь kk и M можно указывать не только для соло товара, но и для full. 3. Цена в базе данных теперь double, а не float

+++ src/main/java/org/intrigger/ultimate_market/utils/ItemStackNotation.java
@@ -3,13 +3,13 @@ package org.intrigger.ultimate_market.utils;
 public class ItemStackNotation {
     public String key;
     public String owner;
-    public float price;
+    public double price;
     public long time;
     public byte[] bytes;
     public int full;
     public int amount;
 
-    public ItemStackNotation(String _key, String _owner, float _price, long _time, byte[] _bytes, int _amount, int _full){
+    public ItemStackNotation(String _key, String _owner, double _price, long _time, byte[] _bytes, int _amount, int _full){
         key = _key;
         owner = _owner;
         price = _price;
